critical changes made
For Assignment 02 changes were made  to the average calories function to return an integer instead of a float

+++ assignments/assignment02/data_structures/Food_utilities.py
@@ -153,7 +153,7 @@ def average_calories(foods):
     for food in foods:
         total += food.calories
         avg += 1
-    avg = total / avg
+    avg = int(total / avg)
     return avg
